+++ src/modules/rolle/api/rolle-with-serviceprovider.response.ts
@@ -1,22 +1,9 @@
-/* eslint-disable max-classes-per-file */
 import { ApiProperty } from '@nestjs/swagger';
 
 import { ServiceProvider } from '../../service-provider/domain/service-provider.js';
 import { Rolle } from '../domain/rolle.js';
 import { RolleResponse } from './rolle.response.js';
-
-class ServiceProviderIdNameResponse {
-    @ApiProperty()
-    public id: string;
-
-    @ApiProperty()
-    public name: string;
-
-    public constructor(serviceProvider: ServiceProvider<true>) {
-        this.id = serviceProvider.id;
-        this.name = serviceProvider.name;
-    }
-}
+import { ServiceProviderIdNameResponse } from './serviceprovider-id-name.response.js';
 
 export class RolleWithServiceProvidersResponse extends RolleResponse {
     @ApiProperty({ type: ServiceProviderIdNameResponse, isArray: true })
